[fix] fix #223 (#242)
동영상 상세 페이지의 동영상 수정/삭제 버튼이 작성자에게만 보이도록 변경합니다.

+++ frontend/js/video/detail.js
@@ -1,36 +1,57 @@
 const detailView = function() {
     const videoId = wootubeCtx.util.getUrlParams('id');
     api.requestVideo(videoId)
-    .then(response => {
-        if (response.status !== 200) {
-            response.json()
-            .then(err => {
-                alert('해당 동영상이 없습니다.');
-                window.location.href = "/"
-            })
-        } else {
-            return response.json()
-        }
-    })
-    .then(data => {
-        detailVideo(data)
-        document.querySelector('#btn-edit').addEventListener('click', () => {
-            location.href = `/video-edit.html?id=${wootubeCtx.util.getUrlParams('id')}`
+        .then(response => {
+            if (response.status !== 200) {
+                response.json()
+                .then(err => {
+                    alert('해당 동영상이 없습니다.');
+                    window.location.href = "/"
+                })
+            } else {
+                return response.json()
+            }
         })
-        
-        document.querySelector('#btn-delete').addEventListener('click', () => {
-            api.deleteVideo(wootubeCtx.util.getUrlParams('id'))
-                .then(response => deleteVideo(response))
-        });
+        .then(data => detailVideo(data))
+        .then(data => {
+            const btnEdit = document.querySelector('#btn-edit')
+            const btnDelete = document.querySelector('#btn-delete')
 
-    })
+            if (btnEdit) {
+                btnEdit.addEventListener('click', () => {
+                    location.href = `/video-edit.html?id=${wootubeCtx.util.getUrlParams('id')}`
+                })
+            }
+
+            if (btnDelete) {
+                btnDelete.addEventListener('click', () => {
+                    api.deleteVideo(wootubeCtx.util.getUrlParams('id'))
+                        .then(response => deleteVideo(response))
+                })
+            }
+        })
 }
 
 const detailVideo = function(data) {
-    addVideoDetailTemplate(data)
+    return new Promise((resolve) => {
+        api.retrieveLoginInfo()
+            .then(res => {
+                if (res.status === 200) {
+                    return res.json()
+                }
+            })
+            .then(json => {
+                if (json) {
+                    addVideoDetailTemplate(data, data.creator.id === json.id)
+                    return resolve()
+                }
+                addVideoDetailTemplate(data, false)
+                return resolve()
+            })
+    })
 }
 
-detailView();
+detailView()
 
 const deleteVideo = function(data) {
     if (data.status === 404) {
